deleted linux config and make files

+++ src/ofxUILabelButton.h
@@ -103,12 +103,10 @@ public:
         float w = labelrect->getWidth(); 
         float pw = rect->getWidth(); 
         
-		labelrect->y = ph*.5 - h*.5; 
-        labelrect->x = pw*.5 - w*.5-padding*.5; 
+		labelrect->y = (int)(ph*.5 - h*.5); 
+        labelrect->x = (int)(pw*.5 - w*.5-padding*.5); 
 		paddedRect->height = rect->height+padding*2.0;  
         paddedRect->width = rect->width+padding*2.0;  
-        
-        
 	}
